CODEONE/DATAMATRIX/MAILMARK/PLESSEY: fix some 32-bit/portability bugs PLESSEY: add options NCR weighted mod-10, hide check digit(s) in HRT test suite: now runnable under MSVC 2019, 2017, 2015, MinGW/MSYS win32/README: update with MSVC 2019 and CMake instructions

+++ backend/common.c
@@ -462,7 +462,7 @@ void debug_test_codeword_dump(struct zint_symbol *symbol, const unsigned char *c
     int i, max = length, cnt_len = 0;
     if (length > 30) { /* 30*3 < errtxt 92 (100 - "Warning ") chars */
         sprintf(symbol->errtxt, "(%d) ", length); /* Place the number of codewords at the front */
-        cnt_len = strlen(symbol->errtxt);
+        cnt_len = (int) strlen(symbol->errtxt);
         max = 30 - (cnt_len + 2) / 3;
     }
     for (i = 0; i < max; i++) {
